+++ docs/conf.py
@@ -69,7 +69,7 @@ author = "Eleanor Smith"
 # the built documents.
 #
 # The short X.Y version.
-version = "0.6.5"
+version = "0.6.6"
 # The full version, including alpha/beta/rc tags.
 release = version